Update left_sidebar-2019.html
Need to start preparing the CFP to be put online. The suggestion this year is to not have separate links for VAST, InfoVis, or SciVis (so I didn't add those), but from the CFP you will link to those. Working on adding the CFP pages next. Sorry if I mess anything up.